+++ src/oloo.js
@@ -1,5 +1,19 @@
 /**
  * @function
+ * @private
+ *
+ * @description
+ * Check if passed item is an object
+ *
+ * @param {*} item The item to be checked
+ * @returns {boolean} The passed item is an object
+ */
+function isObject(item) {
+  return typeof item === 'object';
+}
+
+/**
+ * @function
  * @public
  *
  * @description
@@ -10,7 +24,11 @@
  * @returns {Object} New instance of a combined object
  */
 function oloo(origin, ...sources) {
-  return Object.assign(Object.create(origin), ...sources);
+  if (isObject(origin) && sources.every(isObject)) {
+    return Object.assign(Object.create(origin), ...sources);
+  } else {
+    return null;
+  }
 }
 
 module.exports = oloo;